requirements.txt
Flask==3.1.1
meross-iot==0.4.9.0
asyncio-mqtt==0.16.2
aiohttp==3.12.14
paho-mqtt==2.1.0
python-dotenv==1.1.1
waitress==3.0.2
